requirements.txt
feedparser 
mysql-connector-python 
requests 
textstat 
nltk 
certifi
flask 
pandas
numpy 
scikit-learn 
transformers 
spacy 
requests 
beautifulsoup4 
python-dotenv 
python-dateutil
fastapi 
uvicorn